+++ apps/frontend-react/src/components/DeveloperPanel.module.css
