+++ apps/frontend-react/src/components/transmission/TransmissionConfirmDeleteModal.tsx
