+++ apps/frontend-react/src/components/transmission/TransmissionName.tsx
