+++ apps/frontend-react/src/components/transmission/TransmissionSplitProgress.tsx
