+++ apps/frontend-react/src/components/transmission/TransmissionStemIcon.tsx
