+++ apps/frontend-react/src/components/transmission/TransmissionStemPicker.tsx
