+++ apps/frontend-react/src/components/transmission/TransmissionStemPlayer.tsx
