+++ apps/frontend-react/src/components/transmission/upload/TransmissionUploadForm.tsx
@@ -4,8 +4,8 @@ import { useForm } from "@mantine/form";
 import { useDisclosure } from "@mantine/hooks";
 import { IconMusic, IconPhoto, IconUpload, IconX } from "@tabler/icons-react";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import TransmissionAPI from "../../api/TransmissionAPI";
-import isURLSafe from "../../utils/isURLSafe";
+import TransmissionAPI from "../../../api/TransmissionAPI";
+import isURLSafe from "../../../utils/isURLSafe";
 
 interface FormValues {
   file: File | null;
